+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog], and this project adheres to
 [Semantic Versioning].
 
+## [Unreleased]
+
+### Added
+- Deprecation warnings if rules are specified in YAML format.
+
 ## [2.1.3] - 2020-06-01
 
 ### Fixed
@@ -93,6 +98,7 @@ The format is based on [Keep a Changelog], and this project adheres to
 - The initial release, which evolved from an early prototype specific to a
   certain use case into a more generic command line tool.
 
+[Unreleased]: https://github.com/nixcloud/ip2unix/compare/v2.1.3...HEAD
 [2.1.3]: https://github.com/nixcloud/ip2unix/compare/v2.1.2...v2.1.3
 [2.1.2]: https://github.com/nixcloud/ip2unix/compare/v2.1.1...v2.1.2
 [2.1.1]: https://github.com/nixcloud/ip2unix/compare/v2.1.0...v2.1.1